move assertMock to function for pytest
